requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.3
PyYAML==5.1.2
sqlparse==0.3.0
uritemplate==3.0.0
whitenoise==4.1.4
